pyproject.toml
[tool.poetry]
name = "telegramosintapp"
version = "0.1.0"
description = ""
authors = ["itay.dar <itay.dar@lemonade.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
pynest-api = "^0.3.1"
telethon = "^1.37.0"
pytz = "^2024.2"
black = "^24.10.0"
openai = "^1.51.2"
uvicorn = "^0.31.1"
tiktoken = "^0.8.0"
pandas = "^2.2.3"
tqdm = "^4.66.5"
click = "^8.1.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
